Set an explicit cast to clear a compiler warning.
git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@249 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a

+++ src/UI/FactoryProduction.cpp
@@ -22,7 +22,7 @@ void FactoryProduction::init()
 	size(320, 162);
 
 	// Set up GUI Layout
-	addControl("mProductionGrid", &mProductionGrid, constants::MARGIN, 25);
+	addControl("mProductionGrid", &mProductionGrid, static_cast<float>(constants::MARGIN), 25);
 	mProductionGrid.font(font());
 	mProductionGrid.sheetPath("ui/surface_factory.png");
 	mProductionGrid.size(140, 110);
